--- pages/[feed]/index.vue
@@ -1,7 +1,0 @@
-<script setup lang="ts">
-import Comp from './[page].vue'
-</script>
-
-<template>
-  <Comp />
-</template>
